refactor: extract SessionLibraryEmptyState.swift (#665 slice C FINAL)
Third and final domain-decomposition slice of #665. Extracts:
- enum SessionLibraryEmptyState (with 3 cases + associated computed properties)

Byte-identical move; SHA verified against origin/main.

Reduces SessionLibrary.swift from 268 to 216 lines. Only the primary
SessionLibrary query engine enum remains.

After #665 slices A/B/C: SessionLibrary.swift went from 419 to 216 lines
(~48% reduction), with the file focused solely on the query engine.

Tests: 667 (unchanged).
